Add the main scaffolding for initialization of the keychain
The keychain is created so that it stores the public and private key in local storage.
If there is already a key in local storage, it tries to load that first, but if it fails
it will overwrite them with newly generate keys.

We now need to verify the proper functioning of the keys by creating the specification for encrypt and decrypt.

I had to add a ready function because the used Crypto API is async, and constructor methods are not allowed to be async.